linebot-project/requirements.txt
fastapi
uvicorn[standard]
line-bot-sdk
python-dotenv